evaluation-observe/open-telemetry-instrumentation/requirements.txt
boto3>=1.28.0
opentelemetry-api>=1.19.0
opentelemetry-sdk>=1.19.0
opentelemetry-exporter-otlp>=1.19.0
wrapt>=1.14.0